virt_mshv: separate SNP hypercall handling
Copilot-Session: 73d1c78f-7f6b-4209-87ec-668df6c7547b
